[Lake] Create lake tablet in BE
1. generate lake TabletMetadataPB and write to object storage.
2. refactor convert thrift schema to protobuf schema.
