+++ gensrc/thrift/AgentService.thrift
@@ -52,7 +52,8 @@ enum TStorageFormat {
 
 enum TTabletType {
     TABLET_TYPE_DISK = 0,
-    TABLET_TYPE_MEMORY = 1
+    TABLET_TYPE_MEMORY = 1,
+    TABLET_TYPE_LAKE = 2
 }
 
 struct TCreateTabletReq {
